Stream API now supports acquire / release semantics.
Previously, Streams had two modes (native and texid), this adds a third
mode called "acquired", which allows for copy-free synchronized external
textures in OpenGL and paves the way for Vulkan.

Initially, I attempted to add this functionality directly into Texture.
However, because it requires additional state tracking, I ultimately
decided that the existing Stream API would be a better fit.

This API is probably not necessary for Metal due to its use of shared
ownership semantics.

This has been tested with a new Android sample that will be added in a
subsequent PR.